requirements.txt
python-telegram-bot==22.0
ccxt
python-dotenv
pandas